+++ pydidas/multiprocessing/app_processor_.py
@@ -142,7 +142,9 @@ def app_processor(
                 if _signal is not None:
                     _signal_queue.put(_signal)
                     while not _app.signal_processed_and_can_continue():
-                        time.sleep(0.005)
+                        time.sleep(0.01)
+                    if _results is None:
+                        _results = _app.get_latest_results()
                 _output_queue.put([_arg, _results])
                 _debug_message("Finished computation of item %s" % _arg)
         if not _app_carryon: